feat : utill 함수 분리
